requirements.txt
azure-functions==1.0.8
certifi==2019.11.28
chardet==3.0.4
idna==2.8
joblib==0.14.1
numpy==1.18.1
pandas==1.0.1
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scikit-learn==0.20.0
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8